1 frame
Starting over from scratch with some slightly different settings.

The existing TAS only syncs due to a combination of a lot of
coincidental thread timings and libTAS bugs that have already been
fixed, so there's very little hope for salvaging it.

I'm also going to publish this on GitHub instead of hiding it until it's
done because at this point I'm not too worried about someone stealing
the TAS and claiming it's their own with the level of community
interaction I've had in my progress so far.